apps/backend/requirements.txt
asgiref==3.11.1
attrs==26.1.0
autobahn==25.12.2
Automat==25.4.16
cachetools==7.0.5
cbor2==5.8.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.1
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cron_descriptor==2.0.6
croniter==6.2.2
crontab==1.0.5
cryptography==46.0.5
Django==6.0.3
django-cors-headers==4.9.0
django-filter==25.2
django-redis==6.0.0
django-rq==4.0.1
django-stubs-ext==6.0.1
django-tasks==0.12.0
django-tasks-rq==0.12.0
django-timezone-field==7.2.1
djangorestframework==3.17.0
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
environs==14.6.0
freezegun==1.5.5
gunicorn==25.1.0
httplib2==0.31.2
humanize==4.12.1
hyperlink==21.0.0
idna==3.10
Incremental==24.11.0
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
marshmallow==4.2.2
msgpack==1.1.2
packaging==26.0
pillow==12.1.1
prompt_toolkit==3.0.50
proto-plus==1.27.1
protobuf==6.33.6
psycopg2-binary==2.9.10
py-ubjson==0.16.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.12.1
pyOpenSSL==26.0.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.1
PyYAML==6.0.3
redis==7.3.0
referencing==0.37.0
requests==2.32.4
rpds-py==0.30.0
rq==2.7.0
rq-scheduler==0.14.0
rsa==4.9
service-identity==24.2.0
setuptools==82.0.1
six==1.17.0
sqlparse==0.5.5
tornado==6.5.5
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.13.2
tzdata==2025.1
ujson==5.12.0
uritemplate==4.2.0
urllib3==2.6.3
wcwidth==0.2.13
zope.interface==8.2
